Refactored CMakeLists.txt for better editing and overview Refactored all Defines for Gurobi, TBB, etc into the storm-config file Fixed a missing cast int SymbolicModelAdapter.h Fixed changed iterator structures in SparseMatrix.h Fixed bugs in CuddUtility.cpp where a 64bit shift was executed on a 32bit literal (1 should be 1ull) Fixed a Type Error in graph.h
Former-commit-id: 797b4da2ebecb716c793d0529b7012737e448b2a

+++ src/adapters/SymbolicModelAdapter.h
@@ -206,7 +206,7 @@ private:
 
 		*systemAdd *= *reachableStates;
 		std::cout << "got " << systemAdd->nodeCount() << " nodes" << std::endl;
-		std::cout << "and " << systemAdd->CountMinterm(allRowDecisionDiagramVariables.size() + allColumnDecisionDiagramVariables.size()) << std::endl;
+		std::cout << "and " << systemAdd->CountMinterm(static_cast<int>(allRowDecisionDiagramVariables.size() + allColumnDecisionDiagramVariables.size())) << std::endl;
 	}
 
 	void createIdentityDecisionDiagrams(storm::ir::Program const& program) {
